# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
  "/user/users":
    get:
      description: "Gets all users."
      tags:
        - User
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched all users."
                data: []
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/user/login":
    post:
      description: Logs a user in.
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                usernameOrEmail:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [success]
                    example: "success"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "Successfully logged in."
                      code:
                        type: number
                        enum: [2000]
                      jwtToken:
                        type: string
                        nullable: true
                        example: zqpmxwonceib123
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [error]
                    example: "error"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "Invalid request body."
                      code:
                        type: number
                        enum: [1012]
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [error]
                    example: "error"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "An error was caught. Error: something went wrong."
                      code:
                        type: number
                        enum: [1018]
  "/user/reset-password":
    post:
      description: Sends a password reset email.
      tags:
        - User
      requestBody:
        description: "JWT and the new password."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
            example:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJpYXQiOjE1OTYwNzQwNzcsImV4cCI6MTU5NjA3NzY3N30.5f3b3b3b3b3b3b3b3b3b3b3b3b"
              password: "newpassw0rd"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: true
                userFound: true
                message: "Reset password email sent."
                data: "123objectId321"
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: false
                userFound: false
                message: "Invalid request body."
                data: null
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: false
                userFound: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/user/signup":
    post:
      description: Creates a new user.
      tags:
        - User
      requestBody:
        description: "A username, email and a password."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              username: "username123"
              email: "my@email.com"
              password: "newpassw0rd"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  usernameAvailable:
                    type: boolean
                    example: true
                  emailAvailable:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully created the user."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: "5f3b3b3b3b3b3b3b3b3b3b3b"
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  usernameAvailable:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...